attacker4: single allocation, color selection
-m <memsize>
use a single global memory allocation instead of per-liked list
allocation in prior versions. so the <memsize> is the total allocated
memory regardless of the mlp value. i.e., each list gets a fraction of
the allocated memory.

-e <color>
if selected, the working-set is created using the selected color. the
number of available colors depends on the number of bits in the -b
<bitmask>. multiple colors can be selected by conjunction: -e 1 -e 2

Examples

  $ ./attacker4 -m 48 -l 6
	48KB single buffer, out of which 6 linked lists are created.

  $ ./attacker4 -m 720 -l 6 -e 1
	720KB buffer, 6 lists with color 1 (actual working-set: ~45KB),
      	dram bank bits: 11-14

  $ ./attacker4 -m 16384 -l 6
	16MB buffer, 6 lists

  $ ./attacker4 -m 65536 -l 6 -e 0 -e 15
	64MB buffer, 6 lists with color 0 and 15 (WS: ~8MB).